# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "RECA", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "tags": ["projects"],
        "summary": "Create Project",
        "operationId": "projects_create_project_post_api_v1_projects",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "get": {
        "tags": ["projects"],
        "summary": "List Projects",
        "operationId": "projects_list_projects_get_api_v1_projects",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ProjectStatus" },
                { "type": "null" }
              ],
              "title": "Status"
            }
          },
          {
            "name": "current_stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ProjectStage" },
                { "type": "null" }
              ],
              "title": "Current Stage"
            }
          },
          {
            "name": "project_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ProjectType" },
                { "type": "null" }
              ],
              "title": "Project Type"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "maxLength": 200 },
                { "type": "null" }
              ],
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "desc", "title": "Order" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectListEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Project",
        "operationId": "projects_get_project_get_api_v1_projects_project_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "patch": {
        "tags": ["projects"],
        "summary": "Update Project",
        "operationId": "projects_update_project_patch_api_v1_projects_project_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/overview": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Project Overview",
        "operationId": "projects_get_project_overview_get_api_v1_projects_project_id_overview",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverviewEnvelope"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/archive": {
      "post": {
        "tags": ["projects"],
        "summary": "Archive Project",
        "operationId": "projects_archive_project_post_api_v1_projects_project_id_archive",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/restore": {
      "post": {
        "tags": ["projects"],
        "summary": "Restore Project",
        "operationId": "projects_restore_project_post_api_v1_projects_project_id_restore",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractErrorResponse": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ContractErrorDetail" }
        },
        "type": "object",
        "required": ["error"],
        "title": "ContractErrorResponse"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "discipline": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Discipline"
          },
          "research_direction": {
            "anyOf": [
              { "type": "string", "maxLength": 200 },
              { "type": "null" }
            ],
            "title": "Research Direction"
          },
          "project_type": { "$ref": "#/components/schemas/ProjectType" },
          "current_stage": {
            "$ref": "#/components/schemas/ProjectStage",
            "default": "INTENT"
          },
          "expected_completion_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Expected Completion Date"
          },
          "resource_constraints": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Resource Constraints"
          },
          "ethical_constraints": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Ethical Constraints"
          }
        },
        "type": "object",
        "required": ["name", "project_type"],
        "title": "ProjectCreate"
      },
      "ProjectEnvelope": {
        "properties": {
          "data": { "$ref": "#/components/schemas/ProjectPublic" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "type": "object",
        "required": ["data", "meta"],
        "title": "ProjectEnvelope"
      },
      "ProjectListEnvelope": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ProjectPublic" },
            "type": "array",
            "title": "Data"
          },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "type": "object",
        "required": ["data", "pagination", "meta"],
        "title": "ProjectListEnvelope"
      },
...
      "ProjectOverviewEnvelope": {
        "properties": {
          "data": { "$ref": "#/components/schemas/ProjectOverviewPublic" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "type": "object",
        "required": ["data", "meta"],
        "title": "ProjectOverviewEnvelope"
      },
...
      "ProjectStage": {
        "type": "string",
        "enum": [
          "INTENT",
          "LITERATURE",
          "REVIEW",
          "TOPIC",
          "DATA",
          "ANALYSIS",
          "FIGURE",
          "MANUSCRIPT",
          "EVIDENCE",
          "EXPORT"
        ],
        "title": "ProjectStage"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": ["ACTIVE", "ARCHIVED", "DELETED"],
        "title": "ProjectStatus"
      },
      "ProjectType": {
        "type": "string",
        "enum": ["THESIS", "COURSE", "INNOVATION", "RESEARCH", "DEMO"],
        "title": "ProjectType"
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 200, "minLength": 1 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "discipline": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Discipline"
          },
          "research_direction": {
            "anyOf": [
              { "type": "string", "maxLength": 200 },
              { "type": "null" }
            ],
            "title": "Research Direction"
          },
          "project_type": {
            "anyOf": [
              { "$ref": "#/components/schemas/ProjectType" },
              { "type": "null" }
            ]
          },
          "current_stage": {
            "anyOf": [
              { "$ref": "#/components/schemas/ProjectStage" },
              { "type": "null" }
            ]
          },
          "expected_completion_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Expected Completion Date"
          },
          "resource_constraints": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Resource Constraints"
          },
          "ethical_constraints": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Ethical Constraints"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": { "type": "string", "title": "Request Id" },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "idempotency_replayed": {
            "type": "boolean",
            "title": "Idempotency Replayed",
            "default": false
          }
        },
        "type": "object",
        "required": ["request_id"],
        "title": "ResponseMeta"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}